requirements.txt
requests>=2.26.0
python-decouple>=3.5
pyTelegramBotAPI>=4.2.2